A python code to retrieve trained weights
When I tried to use export the weights of a trained model to use in a C++ program I faced the problem that using tensorflow I was not able to retrieve the trained coefficients. After some search I'have found a way to do that using keras.
This code is based on the tensorflow example https://github.com/tensorflow/models/tree/master/samples/core/get_started.